+++ CHANGELOG.md
@@ -1,13 +1,13 @@
 # Changelog
 
 ## Unreleased changes
-- If an account does not have any pending transactions, then the transaction table is no longer used for tracking next available account nonce but instead the lfb state is used.
+- If an account does not have any non-finalized transactions, then the transaction table is no longer used for tracking next available account nonce.
+  If there are no non-finalized transactions for an account then keep track of the next available account nonce via the last finalized block.
 
 ## 6.2.1
 
 - The account map is now kept solely on disk in a separate lmdb database and it is no longer part of the internal block state database.
   This change results in less memory usage per account and a decrease in the growth of the database.
-
 - Remove V1 GRPC API from the node. This removes configuration options
   `CONCORDIUM_NODE_RPC_SERVER_PORT`, `CONCORDIUM_NODE_RPC_SERVER_ADDRESS`, 
   `CONCORDIUM_NODE_RPC_SERVER_TOKEN`, `CONCORDIUM_NODE_DISABLE_RPC_SERVER_NODE_ENDPOINTS`